requirements.txt
python-telegram-bot==21.10
httpx==0.28.1
openai==1.68.2
python-dotenv==1.0.1
psutil==7.0.0
playwright==1.52.0
duckduckgo-search==7.5.5
pyautogui==0.9.54
pywinauto==0.6.8
pywin32==310
screen-brightness-control==0.24.2
pycaw==20240210
comtypes==1.4.10
pytesseract==0.3.13
Pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.2.3
PyInstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
